+++ recipe/req_sched/sched_ray_trainer.py
@@ -18,13 +18,8 @@ FSDP PPO Trainer with Ray-based single controller.
 This trainer supports model-agonistic model initialization with huggingface
 """
 
-import glob
-import heapq
-import json
-import os
 import uuid
 from collections import defaultdict
-from contextlib import contextmanager
 from copy import deepcopy
 from pprint import pprint
 from typing import Optional
@@ -32,7 +27,6 @@ from typing import Optional
 import numpy as np
 import ray
 import torch
-from codetiming import Timer
 from torch.utils.data import Dataset, Sampler
 from tqdm import tqdm
 
@@ -60,323 +54,10 @@ from verl.utils.debug import marked_timer
 from verl.utils.metric import (
     reduce_metrics,
 )
-from verl.utils.seqlen_balancing import karmarkar_karp
+
+from .ReqScheduler import ReqScheduler, unpad_responses
 
 WorkerType = type[Worker]
-
-
-def unpad_responses(padded_tensor, pad_token_id):
-    if isinstance(pad_token_id, list):
-        # from all worker
-        pid = pad_token_id[0]
-        for worker_pad_token_id in pad_token_id:
-            if worker_pad_token_id != pid:
-                raise ValueError("pad_token_id is not the same across all workers")
-        pad_token_id = pid
-
-    padded_tensor = padded_tensor.cpu()
-    # Convert tensor to list if it's a tensor
-    if isinstance(padded_tensor, torch.Tensor):
-        padded_list = padded_tensor.tolist()
-    else:
-        padded_list = padded_tensor
-
-    # Reconstruct original responses by removing padding tokens
-    unpadded_responses = []
-    for padded_response in padded_list:
-        # Find where padding starts (first occurrence of pad_token_id)
-        try:
-            pad_start_idx = padded_response.index(pad_token_id)
-            # Get only the tokens before padding
-            original_response = padded_response[:pad_start_idx]
-        except ValueError:
-            # No padding found, use the full response
-            original_response = padded_response
-
-        unpadded_responses.append(original_response)
-    return unpadded_responses
-
-
-class ReqScheduler:
-    def __init__(self, config):
-        self.config = config
-
-        # prompt_ids -> len(reponse)
-        self.table: dict[tuple[int], int] = self.load_table()
-
-    def load_table(self):
-        """
-        {
-            "prompts": [
-                [prompt_token_ids_1],
-                [prompt_token_ids_2],
-                ...
-            ],
-            "lengths": [
-                [120, 88, 85, 92, 95, 100, 90, 110],  // prompt 1 对应 sample n 个 response 长度
-                [105, 90, 95, 92, 100, 94, 90, 88],   // prompt 2
-                ...
-            ],
-            "stats": [ // 初始预计算存储，仍可保留便于快速调用
-                {"max": 120, "min": 85, "mean": 97.5, "std": 10.2, "sum": 780},
-                {"max": 105, "min": 88, "mean": 94.3, "std": 5.6, "sum": 754},
-                ...
-            ]
-        }
-        """
-        if self.config.seq_dir is None:
-            return {}
-
-        json_files = glob.glob(os.path.join(self.config.seq_dir, "*.json"))
-        print(f"[ReqScheduler] Found {len(json_files)} JSON files to process")
-
-        ans = {}
-        for json_file in json_files:
-            filename = os.path.basename(json_file)
-            try:
-                with open(json_file) as f:
-                    data = json.load(f)
-
-                print(f"[ReqScheduler] data keys = {data.keys()} in {filename}")
-
-                ps = data["prompts"]
-                ls = data["lengths"]
-                for p, length in zip(ps, ls, strict=False):
-                    p = tuple(p)
-                    if p not in ans:
-                        ans[p] = length
-                print(f"[ReqScheduler] Processed {filename}, found {len(ans)} unique prompts")
-            except Exception as e:
-                print(f"[ReqScheduler] Error processing {filename}: {str(e)}")
-                raise e
-
-        agg = self.config.get("agg", "mean")
-        if agg == "max":
-            ans = {k: max(v) for k, v in ans.items()}
-        elif agg == "min":
-            ans = {k: min(v) for k, v in ans.items()}
-        elif agg == "mean":
-            ans = {k: int(np.mean(v)) for k, v in ans.items()}
-        elif agg == "median":
-            ans = {k: int(np.median(v)) for k, v in ans.items()}
-        elif agg == "sum":
-            ans = {k: sum(v) for k, v in ans.items()}
-        else:
-            raise ValueError(f"Unknown agg {agg}")
-        print(f"[ReqScheduler] Table-Size: {len(ans)=}")
-        return ans
-
-    def lookup_table(self, prompt):
-        if isinstance(prompt, list):
-            prompt = tuple(prompt)
-        assert isinstance(prompt, tuple), f"prompt type {type(prompt)} is not supported"
-        if prompt in self.table:
-            return self.table[prompt]
-        return None
-
-    def update_table(self, raw_prompt_ids, responses):
-        new_table = {}
-        for p, r in zip(raw_prompt_ids, responses, strict=False):
-            p = tuple(p)
-            r = tuple(r)
-            if p not in new_table:
-                new_table[p] = []
-            new_table[p].append(len(r))
-
-        agg = self.config.get("agg", "mean")
-        if agg == "max":
-            new_table = {k: max(v) for k, v in new_table.items()}
-        elif agg == "min":
-            new_table = {k: min(v) for k, v in new_table.items()}
-        elif agg == "mean":
-            new_table = {k: int(np.mean(v)) for k, v in new_table.items()}
-        elif agg == "median":
-            new_table = {k: int(np.median(v)) for k, v in new_table.items()}
-        elif agg == "sum":
-            new_table = {k: sum(v) for k, v in new_table.items()}
-        else:
-            raise ValueError(f"Unknown agg {agg}")
-
-        for k, v in new_table.items():
-            self.table[k] = v
-        print(f"[ReqScheduler] in update_table, Table-Size: {len(self.table)=}")
-
-    def log_seqlen(self, raw_prompt_ids, responses, prefix):
-        print(
-            f"[ReqScheduler] in log_seqlen, {type(raw_prompt_ids)},\
-            {type(responses)}, {len(raw_prompt_ids)}, {len(responses)}"
-        )
-        assert len(raw_prompt_ids) == len(responses), f"{len(raw_prompt_ids)}, {len(responses)}"
-        prompts_dict = {}
-        prompts, response = [], []
-        for p, r in zip(raw_prompt_ids, responses, strict=False):
-            if tuple(p) not in prompts_dict:
-                prompts_dict[tuple(p)] = []
-            prompts_dict[tuple(p)].append(len(r))
-
-        for pid in prompts_dict:
-            prompts.append(list(pid))
-            response.append(prompts_dict[pid])
-
-        log_dir = self.config.log_dir
-        os.makedirs(log_dir, exist_ok=True)
-        data_files = glob.glob(f"{log_dir}/{prefix}_*.json")
-        file_num = len(data_files) + 1
-        output_file = f"{log_dir}/{prefix}_{file_num}.json"
-        with open(output_file, "w") as f:
-            json.dump({"prompts": prompts, "lengths": response}, f)
-
-    def restore_order(
-        self,
-        gen_batch_output: DataProto,
-        reqs_idx,
-        n_samples,
-    ):
-        bs = len(gen_batch_output)
-        assert bs % n_samples == 0, f"bs {bs} must be divisible by n_samples {n_samples}"
-        assert bs // n_samples == len(reqs_idx), f"bs//n_samples {bs // n_samples} != len(reqs_idx) {len(reqs_idx)}"
-        print(f"[ReqScheduler] restore_order, {bs=}, {n_samples=}, {len(reqs_idx)=}")
-        cnt = 0
-        global_idx = [None for _ in range(bs)]
-        group_idx = 0
-        max_id = max(reqs_idx) if reqs_idx else -1
-        while group_idx <= max_id:
-            for i, idx in enumerate(reqs_idx):
-                if idx == group_idx:
-                    start_position = i * n_samples
-                    end_position = start_position + n_samples
-                    global_idx[start_position:end_position] = [j for j in range(cnt, cnt + n_samples)]
-                    cnt += n_samples
-            group_idx += 1
-
-        assert len(global_idx) == bs, f"len(global_idx) {len(global_idx)} != bs {bs}"
-
-        global_idx = torch.tensor(global_idx)
-        gen_batch_output.reorder(global_idx)
-
-    def sched(
-        self,
-        batch_dict: dict,
-        world_size: int,
-        config,
-    ):
-        print(f"[ReqScheduler] sched, {world_size=}, {config=}")
-
-        pre_outlens = []
-        for raw_prompt_ids in batch_dict["raw_prompt_ids"]:
-            outlen = self.lookup_table(raw_prompt_ids)
-            pre_outlens.append(outlen)
-
-        # sched
-        tp_size = config.rollout.tensor_model_parallel_size
-        assert world_size % tp_size == 0, f"world_size {world_size} must be divisible by tp_size {tp_size}"
-        dp_size = world_size // tp_size
-        res = self._sched(pre_outlens, dp_size, tp_size)
-
-        batch_dict["reqs_idx"] = res
-        batch_dict["pre_outlens"] = np.array(pre_outlens, dtype=np.int32)
-
-    def print_stats(self, outlens, res):
-        longest = max(outlens)
-        shortest = min(outlens)
-        avg = np.mean(outlens)
-        std = np.std(outlens)
-        print(f"[ReqScheduler] Stats: {longest=}, {shortest=}, avg: {avg:.2f}, std: {std:.2f}")
-        num_group = np.unique(res)
-        group = [0 for _ in range(len(num_group))]
-        for v in res:
-            group[v] += 1
-        print(f"[ReqScheduler] Group: {group}")
-
-    def _sched(self, outlens, dp_size, tp_size):
-        algo = self.config.algo
-
-        has_none = False
-        for outlen in outlens:
-            if outlen is None:
-                has_none = True
-                break
-
-        agg = self.config.get("agg", "mean")
-        if has_none:
-            print(f"[ReqScheduler] has None, reset {algo} to even_prompt; {agg=}")
-            algo = "even_prompt"
-
-            for i in range(len(outlens)):
-                outlens[i] = -1
-        else:
-            print(f"[ReqScheduler] algo: {algo}, {agg=}")
-
-        method = getattr(self, algo)
-        res = method(outlens, dp_size, tp_size, self.config)
-        self.print_stats(outlens, res)
-        return res
-
-    def dummy(self, outlens, dp_size, tp_size, config):
-        res = [0] * (len(outlens) - 1) + [1]
-        res = np.array(res, dtype=np.int32)
-        return res
-
-    def even_prompt(self, outlens: list[int], dp_size, tp_size, config):
-        num_prompts = len(outlens)
-        if num_prompts == 0:
-            return np.array([], dtype=np.int32)
-        if dp_size <= 0:
-            raise ValueError("dp_size must be a positive integer.")
-
-        base_prompts_per_dp = num_prompts // dp_size
-        remainder_prompts = num_prompts % dp_size
-        res = []
-        for i in range(dp_size):
-            num_in_group = base_prompts_per_dp + (1 if i < remainder_prompts else 0)
-            res.extend([i] * num_in_group)
-        return np.array(res, dtype=np.int32)
-
-    def even_token(self, outlens, dp_size, tp_size, config):
-        prompt_indices = list(range(len(outlens)))
-        sorted_pairs = sorted(zip(outlens, prompt_indices, strict=False), reverse=True)
-        heap = [(0, i) for i in range(dp_size)]
-        heapq.heapify(heap)
-        res = [None] * len(outlens)
-        for token_len, orig_idx in sorted_pairs:
-            total, group = heapq.heappop(heap)
-            res[orig_idx] = group
-            heapq.heappush(heap, (total + token_len, group))
-        return np.array(res, dtype=np.int32)
-
-    def even_token_kk(self, outlens: list[int], dp_size: int, tp_size: int, config):
-        """
-        Schedules requests to balance the total number of tokens per DP group
-        using the Karmarkar-Karp (KK) number partitioning algorithm.
-        """
-        if not outlens:
-            return np.array([], dtype=np.int32)
-
-        print(f"[ReqScheduler] Running Karmarkar-Karp for {len(outlens)} prompts into {dp_size} groups.")
-
-        partitions = karmarkar_karp(
-            seqlen_list=outlens,
-            k_partitions=dp_size,
-            equal_size=False,  # We want to balance sum of tokens, not count of prompts
-        )
-
-        res = [None] * len(outlens)
-        for group_idx, partition_indices in enumerate(partitions):
-            for original_prompt_idx in partition_indices:
-                res[original_prompt_idx] = group_idx
-
-        assert None not in res, "Karmarkar-Karp scheduling failed: not all prompts were assigned a group."
-
-        return np.array(res, dtype=np.int32)
-
-
-@contextmanager
-def _timer(name: str, timing_raw: dict[str, float]):
-    with Timer(name=name, logger=None) as timer:
-        yield
-    if name not in timing_raw:
-        timing_raw[name] = 0
-    timing_raw[name] += timer.last
 
 
 class RaySchedTrainer(RayPPOTrainer):
@@ -412,6 +93,7 @@ class RaySchedTrainer(RayPPOTrainer):
             device_name=device_name,
         )
 
+        # [ReqScheduler] Initialize the request scheduler
         self.req_scheduler = ReqScheduler(
             config=self.config.req_scheduler,
         )
@@ -420,17 +102,18 @@ class RaySchedTrainer(RayPPOTrainer):
         data_source_lst = []
         reward_extra_infos_dict: dict[str, list] = defaultdict(list)
 
+        # Lists to collect samples for the table
         sample_inputs = []
         sample_outputs = []
         sample_scores = []
 
         for test_data in self.val_dataloader:
+            # [ReqScheduler] Schedule the test data
             self.req_scheduler.sched(
                 test_data,
                 self.actor_rollout_wg.world_size,
                 self.config.actor_rollout_ref,
             )
-            print(">> test_data = ", test_data.keys())
             test_batch = DataProto.from_single_dict(test_data)
 
             # repeat test batch
@@ -449,6 +132,7 @@ class RaySchedTrainer(RayPPOTrainer):
             sample_inputs.extend(input_texts)
 
             batch_keys_to_pop = ["input_ids", "attention_mask", "position_ids"]
+            # [ReqScheduler] Pop keys that are not needed for scheduler and generation
             non_tensor_batch_keys_to_pop = ["raw_prompt_ids", "reqs_idx", "pre_outlens"]
             if "multi_modal_data" in test_batch.non_tensor_batch:
                 non_tensor_batch_keys_to_pop.append("multi_modal_data")
@@ -463,6 +147,7 @@ class RaySchedTrainer(RayPPOTrainer):
                 non_tensor_batch_keys=non_tensor_batch_keys_to_pop,
             )
 
+            # [ReqScheduler] Get request indices for  scheduling
             test_reqs_idx = test_gen_batch.non_tensor_batch["reqs_idx"]
             test_gen_batch.meta_info = {
                 "eos_token_id": self.tokenizer.eos_token_id,
@@ -476,7 +161,10 @@ class RaySchedTrainer(RayPPOTrainer):
             test_gen_batch_padded, _ = pad_dataproto_to_divisor(test_gen_batch, self.actor_rollout_wg.world_size)
 
             test_output_gen_batch = self.actor_rollout_wg.generate_sequences(test_gen_batch_padded)
+            # [ReqScheduler] Note that the output is already unpadded in fsdp_workers in our design
+            # [ReqScheduler] So we don't need to unpad it here
 
+            # [ReqScheduler] Restore the order of the generated sequences
             self.req_scheduler.restore_order(
                 test_output_gen_batch, test_reqs_idx, n_samples=self.config.actor_rollout_ref.rollout.val_kwargs.n
             )
@@ -585,6 +273,7 @@ class RaySchedTrainer(RayPPOTrainer):
 
         for epoch in range(self.config.trainer.total_epochs):
             for bs_idx, batch_dict in enumerate(self.train_dataloader):
+                # [ReqScheduler] Schedule the batch for training
                 self.req_scheduler.sched(
                     batch_dict,
                     self.actor_rollout_wg.world_size,
@@ -620,6 +309,7 @@ class RaySchedTrainer(RayPPOTrainer):
                     non_tensor_batch_keys_to_pop.append("tools_kwargs")
                 if "interaction_kwargs" in batch.non_tensor_batch:
                     non_tensor_batch_keys_to_pop.append("interaction_kwargs")
+                # [ReqScheduler] non_tensor_batch_keys for scheduler
                 if "reqs_idx" in batch.non_tensor_batch:
                     non_tensor_batch_keys_to_pop.append("reqs_idx")
                 if "pre_outlens" in batch.non_tensor_batch:
@@ -633,19 +323,13 @@ class RaySchedTrainer(RayPPOTrainer):
 
                 is_last_step = self.global_steps >= self.total_training_steps
 
-                idx = gen_batch.batch["input_ids"]  # (bs, prompt_length)
+                # [ReqScheduler] Get request indices for scheduling and keep ""raw_prompt_ids"" in batch
                 reqs_idx = gen_batch.non_tensor_batch["reqs_idx"]
-                raw_prompt_ids = gen_batch.non_tensor_batch["raw_prompt_ids"]  # (bs, varlen)
-                batch.non_tensor_batch["raw_prompt_ids"] = raw_prompt_ids
-                print(
-                    f"[BATCH INPUT]: \
-                    {idx.shape}, \
-                    {gen_batch.non_tensor_batch.keys()=}, \
-                    raw_prompt_ids = {type(raw_prompt_ids)}"
-                )
+                batch.non_tensor_batch["raw_prompt_ids"] = gen_batch.non_tensor_batch["raw_prompt_ids"]
 
                 with marked_timer("step", timing_raw):
                     with marked_timer("gen", timing_raw, color="red"):
+                        # [ReqScheduler] scheduling for bacth generation
                         gen_batch_output = self.actor_rollout_wg.generate_sequences(gen_batch)
                         timing_raw.update(gen_batch_output.meta_info["timing"])
                         gen_batch_output.meta_info.pop("timing", None)
@@ -664,7 +348,7 @@ class RaySchedTrainer(RayPPOTrainer):
                             batch.batch["reward_baselines"] = reward_baseline_tensor
 
                             del gen_baseline_batch, gen_baseline_output
-
+                    # [ReqScheduler] Restore the order of the generated sequences
                     self.req_scheduler.restore_order(
                         gen_batch_output,
                         reqs_idx,
@@ -677,6 +361,7 @@ class RaySchedTrainer(RayPPOTrainer):
                     batch = batch.repeat(repeat_times=self.config.actor_rollout_ref.rollout.n, interleave=True)
                     batch = batch.union(gen_batch_output)
 
+                    # [ReqScheduler] Log sequence lengths and update the table
                     response = batch.batch["responses"]
                     raw_prompt_ids = batch.non_tensor_batch["raw_prompt_ids"]
 
@@ -685,6 +370,7 @@ class RaySchedTrainer(RayPPOTrainer):
                     dataset = self.config.data.train_files[0].split("/")[-1]
                     prefix = f"{dataset}_{model}_E{epoch}B{bs_idx}_data"
                     unpadded = unpad_responses(response, pad_ids)
+                    # [ReqScheduler] sequence lengths logs can be used for future initialization of scheduler
                     self.req_scheduler.log_seqlen(
                         raw_prompt_ids,
                         unpadded,
@@ -824,7 +510,6 @@ class RaySchedTrainer(RayPPOTrainer):
                     rollout_data_dir = self.config.trainer.get("rollout_data_dir", None)
                     if rollout_data_dir:
                         with marked_timer("dump_rollout_generations", timing_raw, color="green"):
-                            print(batch.batch.keys())
                             inputs = self.tokenizer.batch_decode(batch.batch["prompts"], skip_special_tokens=True)
                             outputs = self.tokenizer.batch_decode(batch.batch["responses"], skip_special_tokens=True)
                             scores = batch.batch["token_level_scores"].sum(-1).cpu().tolist()
