+++ verl/protocol.py
@@ -84,6 +84,8 @@ def pad_dataproto_to_divisor(data: "DataProto", size_divisor: int):
         remaining_pad = pad_size
         while remaining_pad > 0:
             take_size = min(remaining_pad, len(data))
+
+            # [ReqScheduler] if ReqSchduler is enabled, the req_idx of padding data should be -1
             if "reqs_idx" in data.non_tensor_batch:
                 slice_data = data[:take_size]
                 if slice_data.non_tensor_batch is None:
